+++ packages/@storefront/core/test/unit/core/url-beautifier/details/parser.ts
@@ -5,7 +5,17 @@ import suite, { refinement } from '../../../_suite';
 suite('DetailsUrlParser', ({ expect }) => {
   let parser: DetailsUrlParser;
 
-  beforeEach(() => parser = new DetailsUrlParser(<any>{ config: { refinementMapping: [] } }));
+  beforeEach(() => parser = new DetailsUrlParser(<any>{
+    config: {
+      refinementMapping: [],
+      details: {
+        params: {
+          area: 'area',
+          collection: 'collection',
+        },
+      },
+    },
+  }));
 
   it('should extend UrlParser', () => {
     expect(parser).to.be.an.instanceOf(UrlParser);
@@ -95,6 +105,84 @@ suite('DetailsUrlParser', ({ expect }) => {
 
     expect(parsed.data).to.eql(expectedDetail.data);
     expect(parsed.variants).to.eql(expectedDetail.variants);
+  });
+
+  it('should extract "area" query string parameter', () => {
+    const url = 'dress/293014?area=Staging';
+    const expectedDetail = {
+      variants: [],
+      area: 'Staging',
+      data: {
+        id: '293014',
+        title: 'dress',
+      },
+      id: '293014',
+    };
+
+    expect(parser.parse(url)).to.eql(expectedDetail);
+  });
+
+  it('should extract "collection" query string parameter', () => {
+    const url = 'dress/293014?collection=StagingCollection';
+    const expectedDetail = {
+      variants: [],
+      collection: 'StagingCollection',
+      data: {
+        id: '293014',
+        title: 'dress',
+      },
+      id: '293014',
+    };
+
+    expect(parser.parse(url)).to.eql(expectedDetail);
+  });
+
+  it('should extract "myCustomArea" query string parameter', () => {
+    parser.config.details.params.area = 'myCustomArea';
+    const url = 'dress/293014?myCustomArea=Staging';
+    const expectedDetail = {
+      variants: [],
+      area: 'Staging',
+      data: {
+        id: '293014',
+        title: 'dress',
+      },
+      id: '293014',
+    };
+
+    expect(parser.parse(url)).to.eql(expectedDetail);
+  });
+
+  it('should extract "myCustomCollection" query string parameter', () => {
+    parser.config.details.params.collection = 'myCustomCollection';
+    const url = 'dress/293014?myCustomCollection=StagingCollection';
+    const expectedDetail = {
+      variants: [],
+      collection: 'StagingCollection',
+      data: {
+        id: '293014',
+        title: 'dress',
+      },
+      id: '293014',
+    };
+
+    expect(parser.parse(url)).to.eql(expectedDetail);
+  });
+
+  it('should not extract "area" nor "collection" query string parameters', () => {
+    parser.config.details.params.area = '';
+    parser.config.details.params.collection = '';
+    const url = 'dress/293014?area=Staging&collection=StagingCollection';
+    const expectedDetail = {
+      variants: [],
+      data: {
+        id: '293014',
+        title: 'dress',
+      },
+      id: '293014',
+    };
+
+    expect(parser.parse(url)).to.eql(expectedDetail);
   });
 
   describe('error states', () => {