+++ packages/@storefront/core/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] [minor]
+### Added
+- SF-1275: Create configuration option to disable GbTracker beaconing.
+
 ## [2.5.2] - 2019-02-11
 ### Changed
 - SF-1279: Modify autocomplete service's `hasActiveSuggestion` method to take into account whether