+++ packages/@storefront/core/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] [minor]
+### Changed
+- Update `@storefront/flux-capacitor` to 1.72.0.
+
 ### Added
 - SF-1125: Added option for parsing `collection` and `area` through URL on details pages via configurable key-names.
   - Where 'area' and 'collection' are the default key-names for `area` and `collection`, respectively.