+++ packages/@storefront/core/CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] [minor]
+## [2.7.0] - 2019-02-21
+### Changed
+- Update `@storefront/flux-capacitor` to 1.72.0.
+
 ### Added
 - SF-1125: Added option for parsing `collection` and `area` through URL on details pages via configurable key-names.
   - Where 'area' and 'collection' are the default key-names for `area` and `collection`, respectively.